requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
requests==2.32.5
python-dotenv==1.0.1
beautifulsoup4==4.14.2
anthropic==0.71.0
apify-client==2.2.1
aiohttp==3.13.1
python-multipart==0.0.17